@@ -27,7 +27,7 @@ package com.heinrichreimersoftware.materialintro.view;
 import android.content.Context;
 import android.os.Bundle;
 import android.os.Parcelable;
-import android.support.v4.view.ViewPager;
+import androidx.viewpager.widget.ViewPager;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 
@@ -69,12 +69,20 @@ public class SwipeBlockableViewPager extends ViewPager {
 
     @Override
     public boolean onTouchEvent(MotionEvent event) {
-        return handleTouchEvent(event) && super.onTouchEvent(event);
+        try {
+            return handleTouchEvent(event) && super.onTouchEvent(event);
+        } catch (IllegalArgumentException exception) {
+            return false;
+        }
     }
 
     @Override
     public boolean onInterceptTouchEvent(MotionEvent event) {
-        return handleTouchEvent(event) && super.onInterceptTouchEvent(event);
+        try {
+            return handleTouchEvent(event) && super.onInterceptTouchEvent(event);
+        } catch (IllegalArgumentException exception) {
+            return false;
+        }
     }
 
     private boolean handleTouchEvent(MotionEvent event) {
